chore: move helpers folder into app folder and update import path in package.json
