+++ app/helpers/crop_scan_helper.ts
